[gl] extracted some common GPU rendering code
